ci: rebalance builder end-to-end shards
